Refactor creating lib/github/nippou/commands.rb
